+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## v2.0.1
+
+### Fixed
+
+- Disabled all packages iteration for cleaning after **single** package installation or updating
+
+### Added
+
+- Plugin method `::cleanupAllPackages` for batch cleaning (can be executed manually using composer scripts, e.g.: `{"scripts": {"post-update-cmd": "AvtoDev\\Composer\\Cleanup\\Plugin::cleanupAllPackages"}}`)
+
+### Changed
+
+- Output messages format
+
 ## v2.0.0
 
 ### Changed